# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Game backend API với PlayFab integration, anti-cheat validation, và talent system",
        "title": "Hollow Wilds Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@gamefeel.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.1.0"
    },
    "host": "gamefeel-backend.fly.dev",
    "basePath": "/api/v1",
    "paths": {
...
        "/analytics/events": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Submit a batch of analytics events",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HollowWilds"
                ],
                "summary": "Track HW Events",
                "parameters": [
                    {
                        "description": "Analytics events",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AnalyticsEventsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Results",
                        "schema": {
                            "$ref": "#/definitions/models.AnalyticsEventsResponse"
                        }
                    }
                }
            }
        },
        "/auth/hw/login": {
            "post": {
                "description": "Authenticate player with PlayFab ticket and get HW session JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HollowWilds"
                ],
                "summary": "Hollow Wilds Login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "PlayFab ID",
                        "name": "X-PlayFab-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Login request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.HollowWildsLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "schema": {
                            "$ref": "#/definitions/models.HollowWildsAuthResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/models.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/models.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/auth/logout": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Revoke tokens and logout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HollowWilds"
                ],
                "summary": "HW Logout",
                "parameters": [
                    {
                        "description": "Optional refresh token to revoke",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.RefreshTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "Get a new JWT using a refresh token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HollowWilds"
                ],
                "summary": "Refresh HW Token",
                "parameters": [
                    {
                        "description": "Refresh request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RefreshTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "New token",
                        "schema": {
                            "$ref": "#/definitions/models.RefreshTokenResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid refresh token",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/models.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/player/save": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the full game state for the player",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HollowWilds"
                ],
                "summary": "Load HW Game",
                "responses": {
                    "200": {
                        "description": "Game save data",
                        "schema": {
                            "$ref": "#/definitions/models.LoadGameResponse"
                        }
                    },
                    "404": {
                        "description": "Save not found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/models.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Persist the game state with version control",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HollowWilds"
                ],
                "summary": "Save HW Game",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Expected current version",
                        "name": "version",
                        "in": "query"
                    },
                    {
                        "description": "Save data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SaveGameRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Save confirmed",
                        "schema": {
                            "$ref": "#/definitions/models.SaveGameResponse"
                        }
                    },
                    "409": {
                        "description": "Version conflict",
                        "schema": {
                            "$ref": "#/definitions/models.VersionConflictError"
                        }
                    }
                }
            }
        },
        "/player/save/backup": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Manually trigger a save backup",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HollowWilds"
                ],
                "summary": "Create HW Backup",
                "responses": {
                    "200": {
                        "description": "Backup created",
                        "schema": {
                            "$ref": "#/definitions/models.BackupResponse"
                        }
                    }
                }
            }
        },
        "/player/save/backups": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get list of all save backups for the player",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HollowWilds"
                ],
                "summary": "List HW Backups",
                "responses": {
                    "200": {
                        "description": "Backup list",
                        "schema": {
                            "$ref": "#/definitions/models.BackupListResponse"
                        }
                    }
                }
            }
        },
        "/player/save/restore": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Restore game save data from a selected backup",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HollowWilds"
                ],
                "summary": "Restore HW Backup",
                "parameters": [
                    {
                        "description": "Restore request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RestoreRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restore confirmed",
                        "schema": {
                            "$ref": "#/definitions/models.SaveGameResponse"
                        }
                    },
                    "404": {
                        "description": "Backup not found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/models.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/models.APIError"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "models.AnalyticsEvent": {
            "type": "object",
            "required": [
                "event_name",
                "timestamp"
            ],
            "properties": {
                "event_name": {
                    "type": "string"
                },
                "payload": {
                    "type": "object",
                    "additionalProperties": true
                },
                "session_id": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "user_id": {}
            }
        },
        "models.AnalyticsEventsRequest": {
            "type": "object",
            "required": [
                "events"
            ],
            "properties": {
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AnalyticsEvent"
                    }
                }
            }
        },
        "models.AnalyticsEventsResponse": {
            "type": "object",
            "properties": {
                "accepted": {
                    "type": "integer"
                },
                "rejected": {
                    "type": "integer"
                }
            }
        },
...
        "models.BackupListResponse": {
            "type": "object",
            "properties": {
                "backups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BackupInfo"
                    }
                }
            }
        },
        "models.BackupResponse": {
            "type": "object",
            "properties": {
                "backup_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "models.HollowWildsAuthResponse": {
            "type": "object",
            "properties": {
                "expires_in": {
                    "type": "integer"
                },
                "player_id": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "models.HollowWildsLoginRequest": {
            "type": "object",
            "required": [
                "playfab_session_ticket"
            ],
            "properties": {
                "playfab_session_ticket": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
...
        "models.LoadGameResponse": {
            "type": "object",
            "properties": {
                "base": {
                    "$ref": "#/definitions/models.BaseData"
                },
                "discovered_pois": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "inventory": {
                    "$ref": "#/definitions/models.InventoryData"
                },
                "player": {
                    "$ref": "#/definitions/models.PlayerState"
                },
                "player_id": {
                    "type": "string"
                },
                "quest_flags": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "boolean"
                    }
                },
                "save_version": {
                    "type": "integer"
                },
                "sebilah": {
                    "$ref": "#/definitions/models.SebilahData"
                },
                "updated_at": {
                    "type": "string"
                },
                "world": {
                    "$ref": "#/definitions/models.WorldData"
                }
            }
        },
...
        "models.RefreshTokenRequest": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "models.RefreshTokenResponse": {
            "type": "object",
            "properties": {
                "expires_in": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "models.RestoreRequest": {
            "type": "object",
            "required": [
                "backup_id"
            ],
            "properties": {
                "backup_id": {
                    "type": "string"
                }
            }
        },
        "models.SaveGameRequest": {
            "type": "object",
            "properties": {
                "base": {
                    "$ref": "#/definitions/models.BaseData"
                },
                "discovered_pois": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "inventory": {
                    "$ref": "#/definitions/models.InventoryData"
                },
                "player": {
                    "$ref": "#/definitions/models.PlayerState"
                },
                "quest_flags": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "boolean"
                    }
                },
                "sebilah": {
                    "$ref": "#/definitions/models.SebilahData"
                },
                "world": {
                    "$ref": "#/definitions/models.WorldData"
                }
            }
        },
        "models.SaveGameResponse": {
            "type": "object",
            "properties": {
                "save_version": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.VersionConflictError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "server_version": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token format: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "PlayFabToken": {
            "description": "PlayFab session token for authentication",
            "type": "apiKey",
            "name": "X-PlayFab-SessionToken",
            "in": "header"
        }
    }
}